requirements.txt
django
djangorestframework
djangorestframework-jsonapi
django-rest-swagger
drf-nested-routers
django-cors-headers
channels
faker